Need further testing on tool number changes and work offset changes, but appears to work.
Need to add an option to re-write tool changes to accomodate for side-mount or dual arm tool changers that need to prep the next tool.  Basically something to add them in if they aren't added automatically.

Adding options to allow user to turn on self edit of the textbox and also trying to seek by tool and attach to buttons.

I am also going to have a log file generated to show the history of changes made to the user so they can check themselves if things were done how they wanted vs. having to check through thousands of lines of text.